renames main lib file
